# server/docs/swagger.yaml
definitions:
  common.CustomTime:
    properties:
      time.Time:
        type: string
    type: object
...
  petTodo.PetTodoInfo:
    properties:
      color:
        description: 背景颜色
        type: integer
      complete:
        description: 是否完成，0：否，1：是
        type: boolean
      createdAt:
        description: 创建时间
        type: string
      createdBy:
        description: 创建人
        type: integer
      familyId:
        description: 家庭ID
        type: integer
      id:
        description: 主键ID
        type: integer
      petId:
        description: 宠物ID
        type: integer
      remark:
        description: 备注
        type: string
      remind:
        description: 是否提醒， 0：否，1：是
        type: boolean
      remindTime:
        allOf:
        - $ref: '#/definitions/common.CustomTime'
        description: 提醒时间
      title:
        description: 标题
        type: string
      todoTime:
        allOf:
        - $ref: '#/definitions/common.CustomTime'
        description: 待办时间
      type:
        description: 提醒类型， 0：日常记录， 1：待办事项
        type: integer
      updatedAt:
        description: 更新时间
        type: string
      updatedBy:
        description: 更新人
        type: integer
    type: object
  petTodo.SimplePetTodoInfo:
    properties:
      color:
        description: 背景颜色
        type: integer
      complete:
        description: 是否完成，0：否，1：是
        type: boolean
      id:
        description: 主键ID
        type: integer
      petId:
        description: 宠物ID
        type: integer
      remark:
        description: 备注
        type: string
      remind:
        description: 是否提醒， 0：否，1：是
        type: boolean
      remindTime:
        allOf:
        - $ref: '#/definitions/common.CustomTime'
        description: 提醒时间
      title:
        description: 标题
        type: string
      todoTime:
        allOf:
        - $ref: '#/definitions/common.CustomTime'
        description: 待办时间
      type:
        description: 提醒类型， 0：日常记录， 1：待办事项
        type: integer
    type: object
...
  request.PetTodoInfo:
    properties:
      color:
        description: 背景颜色
        type: integer
      complete:
        description: 是否完成，0：否，1：是
        type: boolean
      petId:
        description: 宠物ID
        type: integer
      remark:
        description: 备注
        type: string
      remind:
        description: 是否提醒， 0：否，1：是
        type: boolean
      remindDate:
        description: 提醒日期
        items:
          type: string
        type: array
      remindTime:
        description: 提醒时间
        type: string
      title:
        description: 标题
        type: string
      todoTime:
        allOf:
        - $ref: '#/definitions/common.CustomTime'
        description: 待办时间
      type:
        description: 提醒类型， 0：日常记录， 1：待办事项
        type: integer
    type: object
  request.UserInfo:
    properties:
      avatar:
        description: 用户头像url
        type: string
      id:
        description: 用户ID
        type: integer
      isAdmin:
        description: 是否为管理员
        type: boolean
      nickname:
        description: 昵称
        type: string
    type: object
...
  response.PageResult:
    properties:
      list: {}
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
    type: object
...
  response.PetTodoInfoResponse:
    properties:
      petTodos:
        $ref: '#/definitions/petTodo.PetTodoInfo'
    type: object
  response.Response:
    properties:
      code:
        type: integer
      data: {}
      msg:
        type: string
    type: object
...
info:
  contact: {}
  title: Upet API接口文档
  version: v2.7.0
paths:
...
  /pet-todo:
    post:
      consumes:
      - application/json
      parameters:
      - description: 宠物待办信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.PetTodoInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 创建宠物待办
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 创建宠物待办
      tags:
      - PetTodo
  /pet-todo/:petTodoID:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 宠物待办ID
        in: path
        name: petTodoID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 删除宠物待办
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 删除宠物待办
      tags:
      - PetTodo
    get:
      consumes:
      - application/json
      parameters:
      - description: 宠物待办ID
        in: path
        name: petTodoID
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 获取单一宠物信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PetTodoInfoResponse'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取单一宠物待办信息
      tags:
      - PetTodo
    put:
      consumes:
      - application/json
      parameters:
      - description: 宠物待办ID
        in: path
        name: petTodoID
        required: true
        type: integer
      - description: 宠物待办信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/petTodo.PetTodoInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 更新宠物待办信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新宠物待办信息
      tags:
      - PetTodo
  /pet-todos:
    get:
      consumes:
      - application/json
      parameters:
      - description: 日期
        in: query
        name: date
        type: string
      - description: 宠物id
        in: query
        name: petId
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取权限客户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取所有宠物待办列表
      tags:
      - PetTodo
  /pet-todos/:petTodoID/complete:
    put:
      consumes:
      - application/json
      parameters:
      - description: 宠物待办ID
        in: path
        name: petTodoID
        required: true
        type: integer
      - description: 宠物待办信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/petTodo.PetTodoInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 更新宠物待办信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 更新宠物待办信息
      tags:
      - PetTodo
  /pet-todos/infos:
    get:
      consumes:
      - application/json
      parameters:
      - description: 内容
        in: query
        name: content
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 查询宠物待办信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/petTodo.SimplePetTodoInfo'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 搜索宠物待办信息
      tags:
      - PetTodo
  /pet-todos/mark:
    post:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: 分页获取权限客户列表,返回包括列表,总数,页码,每页数量
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/response.PageResult'
                msg:
                  type: string
              type: object
      security:
      - ApiKeyAuth: []
      summary: 获取指定日期中宠物待办标记
      tags:
      - PetTodo
...
  /user:
    put:
      consumes:
      - application/json
      parameters:
      - description: 用户信息
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.UserInfo'
      produces:
      - application/json
      responses:
        "200":
          description: 用户信息保存结果
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                msg:
                  type: string
              type: object
      summary: 完善用户信息
      tags:
      - User
  /user/:id:
    get:
      parameters:
      - description: 用户id
        in: path
        name: id
        required: true
        type: integer
      responses:
        "200":
          description: 用户信息
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                ' msg':
                  type: string
                data:
                  $ref: '#/definitions/request.UserInfo'
              type: object
      summary: 查询用户信息
      tags:
      - User
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"